requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
ultralytics==8.3.232
openvino==2025.0.0
opencv-python-headless==4.11.0.86
pillow==11.1.0
numpy==2.1.1
